Update salt state to update the mysql conf file and user grants
There was an issue with default my.cnf file which was binding only
to 127.0.0.1 which would make the mysql not accessible from outside.
So this patch updates the modified version of my.cnf file which would
make is bind to 0.0.0.0

Also the user webapp created by the state, lacked the proper permissions
and grants on databases. This patch updates that part, providing access
from everywhere and all grants on all databases.

Ideally, we should restrict the permissions, but for now this will do. I will
come back to this later change it to only required database.

Signed-off-by: M S Vishwanath Bhat <msvbhat@gmail.com>